% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_correspondence.R
\name{areacorrespondence}
\alias{areacorrespondence}
\title{Construct an area-weighted correspondence between two digital boundary files}
\usage{
areacorrespondence(from, to, from.ID, to.ID, filename = NULL)
}
\arguments{
\item{from}{The simple feature geography you wish to correspond from}

\item{to}{The simple feature geography you wish to correspond to}

\item{from.ID}{The column name that identifies the areas in the from geography}

\item{to.ID}{The column name that identifies the areas in the to geography}

\item{filename}{(optional) A csv filename to write the correspondence to when done}
}
\description{
Construct an area-weighted correspondence between two digital boundary files
}
\examples{
\dontrun{
  ASGS11_SA2 <- read_sf("extdata/ASGS2011/SA2_2011_AUST.shp")
  ASGC06_SLA <- read_sf("extdata/ASGC2006/SLA06aAUST.shp")
  areacorrespondence(ASGC06_SLA, ASGS11_SA2, "SLA_CODE06", "SA2_MAIN11")
}

}
