% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correspond.R
\name{correspond}
\alias{correspond}
\title{Convert data from one geography to another}
\usage{
correspond(data, correspondence, data.by, correspondence.from,
  correspondence.to, correspondence.weight)
}
\arguments{
\item{data}{The data you wish to convery}

\item{correspondence}{The correspondence}

\item{data.by}{The column name that identifies the area IDs in your data}

\item{correspondence.from}{The column name that identifies the from areas in the correspondence}

\item{correspondence.to}{The column name that identifies the to areas in the correspondence}

\item{correspondence.weight}{The column name that identifies the weights in the correspondence}
}
\description{
Convert data from one geography to another
}
